requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
yt-dlp>=2024.12.23
